+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add getopt to parse program options
 - Allow NMP, IID, and LMR at shallower depth
 - Set NMP R to 3
+- Save upper and lower bounds in transposition table
 
 
 ## [0.3.0] - 2017-10-11